requirements.txt
asgiref==3.2.7
Django==3.0.6
mysqlclient==1.4.6
paho-mqtt==1.5.0
PyMySQL==0.9.3
pytz==2020.1
simplejson==3.17.0
sqlparse==0.3.1
